scraping/pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Scrape Shopify products from websites"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "huggingface-hub<0.20",
    "numpy<2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers==2.2.2",
    "questions", 
    "psycopg[binary]>=3.1.8",
    "torch==2.1.0",

    "tqdm>=4.67.1",
    "transformers==4.35.2",
    "urllib3>=2.5.0",
    "brotli>=1.0.9",
    "bleach>=6.3.0",
]
